+++ app/tc2026/rijndael_enc/rijndael_enc_libc.cc
@@ -5,28 +5,28 @@ int rijndael_enc_toupper(int c) {
     return c;
 }
 
-unsigned long rijndael_enc_fread(void *ptr, unsigned long size, unsigned long count, struct rijndael_enc_FILE *stream) {
+unsigned int rijndael_enc_fread(void *ptr, unsigned int size, unsigned int count, struct rijndael_enc_FILE *stream) {
     unsigned i = stream->cur_pos, i2 = 0;
-    unsigned long number_of_chars_to_read =
+    unsigned int number_of_chars_to_read =
         stream->size - stream->cur_pos >= size * count ? size * count : stream->size - stream->cur_pos;
     while (i < stream->cur_pos + number_of_chars_to_read) ((unsigned char *)ptr)[i2++] = stream->data[i++];
     stream->cur_pos += number_of_chars_to_read;
     return number_of_chars_to_read;
 }
 
-unsigned long rijndael_enc_fwrite(const void *ptr,
-                                  unsigned long size,
-                                  unsigned long count,
-                                  struct rijndael_enc_FILE *stream) {
+unsigned int rijndael_enc_fwrite(const void *ptr,
+                                 unsigned int size,
+                                 unsigned int count,
+                                 struct rijndael_enc_FILE *stream) {
     unsigned i = stream->cur_pos, i2 = 0;
-    unsigned long number_of_chars_to_write =
+    unsigned int number_of_chars_to_write =
         stream->size - stream->cur_pos >= size * count ? size * count : stream->size - stream->cur_pos;
     while (i < stream->cur_pos + number_of_chars_to_write) stream->data[i++] = ((unsigned char *)ptr)[i2++];
     stream->cur_pos += number_of_chars_to_write;
     return number_of_chars_to_write;
 }
 
-int rijndael_enc_fseek(struct rijndael_enc_FILE *stream, long int offset, RijndaelOrigin origin) {
+int rijndael_enc_fseek(struct rijndael_enc_FILE *stream, int offset, RijndaelOrigin origin) {
     if (origin == RIJNDAEL_ENC_SEEK_SET) {
         stream->cur_pos = offset;
         return 0;
